refactor(storage_impl): Move Conversion impl for MCA to storage_impl crate
