+++ crates/hyperswitch_domain_models/src/merchant_connector_account.rs
@@ -3,12 +3,11 @@ use std::collections::HashMap;
 
 use common_utils::{
     crypto::Encryptable,
-    date_time,
     encryption::Encryption,
-    errors::{CustomResult, ValidationError},
+    errors::CustomResult,
     ext_traits::{StringExt, ValueExt},
-    id_type, pii, type_name,
-    types::keymanager::{Identifier, KeyManagerState, ToEncryptable},
+    id_type, pii,
+    types::keymanager::ToEncryptable,
 };
 #[cfg(feature = "v2")]
 use diesel_models::merchant_connector_account::{
@@ -16,24 +15,17 @@ use diesel_models::merchant_connector_account::{
     MerchantConnectorAccountFeatureMetadata as DieselMerchantConnectorAccountFeatureMetadata,
     RevenueRecoveryMetadata as DieselRevenueRecoveryMetadata,
 };
-use diesel_models::{
-    enums,
-    merchant_connector_account::{self as storage, MerchantConnectorAccountUpdateInternal},
-};
+use diesel_models::{enums, merchant_connector_account::MerchantConnectorAccountUpdateInternal};
 use error_stack::ResultExt;
-use hyperswitch_masking::{PeekInterface, Secret};
+#[cfg(feature = "v2")]
+use hyperswitch_masking::PeekInterface;
+use hyperswitch_masking::Secret;
 use rustc_hash::FxHashMap;
 use serde_json::Value;
 
-use super::behaviour;
 #[cfg(feature = "v2")]
 use crate::errors::api_error_response;
-use crate::{
-    mandates::CommonMandateReference,
-    merchant_key_store::MerchantKeyStore,
-    router_data,
-    type_encryption::{crypto_operation, CryptoOperation},
-};
+use crate::{mandates::CommonMandateReference, merchant_key_store::MerchantKeyStore, router_data};
 
 #[cfg(feature = "v1")]
 #[derive(Clone, Debug, router_derive::ToEncryption)]
@@ -525,412 +517,6 @@ pub enum MerchantConnectorAccountUpdate {
     },
 }
 
-#[cfg(feature = "v1")]
-#[async_trait::async_trait]
-impl behaviour::Conversion for MerchantConnectorAccount {
-    type DstType = storage::MerchantConnectorAccount;
-    type NewDstType = storage::MerchantConnectorAccountNew;
-
-    async fn convert(self) -> CustomResult<Self::DstType, ValidationError> {
-        Ok(storage::MerchantConnectorAccount {
-            merchant_id: self.merchant_id,
-            connector_name: self.connector_name,
-            connector_account_details: self.connector_account_details.into(),
-            test_mode: self.test_mode,
-            disabled: self.disabled,
-            merchant_connector_id: self.merchant_connector_id.clone(),
-            id: Some(self.merchant_connector_id),
-            payment_methods_enabled: self.payment_methods_enabled,
-            connector_type: self.connector_type,
-            metadata: self.metadata,
-            frm_configs: None,
-            frm_config: self.frm_configs,
-            business_country: self.business_country,
-            business_label: self.business_label,
-            connector_label: self.connector_label,
-            business_sub_label: self.business_sub_label,
-            created_at: self.created_at,
-            modified_at: self.modified_at,
-            connector_webhook_details: self.connector_webhook_details,
-            profile_id: Some(self.profile_id),
-            applepay_verified_domains: self.applepay_verified_domains,
-            pm_auth_config: self.pm_auth_config,
-            status: self.status,
-            connector_wallets_details: self.connector_wallets_details.map(Encryption::from),
-            additional_merchant_data: self.additional_merchant_data.map(|data| data.into()),
-            version: self.version,
-            connector_webhook_registration_details: self.connector_webhook_registration_details,
-        })
-    }
-
-    async fn convert_back(
-        state: &KeyManagerState,
-        other: Self::DstType,
-        key: &Secret<Vec<u8>>,
-        _key_manager_identifier: Identifier,
-    ) -> CustomResult<Self, ValidationError> {
-        let identifier = Identifier::Merchant(other.merchant_id.clone());
-        let decrypted_data = crypto_operation(
-            state,
-            type_name!(Self::DstType),
-            CryptoOperation::BatchDecrypt(EncryptedMerchantConnectorAccount::to_encryptable(
-                EncryptedMerchantConnectorAccount {
-                    connector_account_details: other.connector_account_details,
-                    additional_merchant_data: other.additional_merchant_data,
-                    connector_wallets_details: other.connector_wallets_details,
-                },
-            )),
-            identifier.clone(),
-            key.peek(),
-        )
-        .await
-        .and_then(|val| val.try_into_batchoperation())
-        .change_context(ValidationError::InvalidValue {
-            message: "Failed while decrypting connector account details".to_string(),
-        })?;
-
-        let decrypted_data = EncryptedMerchantConnectorAccount::from_encryptable(decrypted_data)
-            .change_context(ValidationError::InvalidValue {
-                message: "Failed while decrypting connector account details".to_string(),
-            })?;
-
-        Ok(Self {
-            merchant_id: other.merchant_id,
-            connector_name: other.connector_name,
-            connector_account_details: decrypted_data.connector_account_details,
-            test_mode: other.test_mode,
-            disabled: other.disabled,
-            merchant_connector_id: other.merchant_connector_id,
-            payment_methods_enabled: other.payment_methods_enabled,
-            connector_type: other.connector_type,
-            metadata: other.metadata,
-
-            frm_configs: other.frm_config,
-            business_country: other.business_country,
-            business_label: other.business_label,
-            connector_label: other.connector_label,
-            business_sub_label: other.business_sub_label,
-            created_at: other.created_at,
-            modified_at: other.modified_at,
-            connector_webhook_details: other.connector_webhook_details,
-            profile_id: other
-                .profile_id
-                .ok_or(ValidationError::MissingRequiredField {
-                    field_name: "profile_id".to_string(),
-                })?,
-            applepay_verified_domains: other.applepay_verified_domains,
-            pm_auth_config: other.pm_auth_config,
-            status: other.status,
-            connector_wallets_details: decrypted_data.connector_wallets_details,
-            additional_merchant_data: decrypted_data.additional_merchant_data,
-            version: other.version,
-            connector_webhook_registration_details: other.connector_webhook_registration_details,
-        })
-    }
-
-    async fn construct_new(self) -> CustomResult<Self::NewDstType, ValidationError> {
-        let now = date_time::now();
-        Ok(Self::NewDstType {
-            merchant_id: Some(self.merchant_id),
-            connector_name: Some(self.connector_name),
-            connector_account_details: Some(self.connector_account_details.into()),
-            test_mode: self.test_mode,
-            disabled: self.disabled,
-            merchant_connector_id: self.merchant_connector_id.clone(),
-            id: Some(self.merchant_connector_id),
-            payment_methods_enabled: self.payment_methods_enabled,
-            connector_type: Some(self.connector_type),
-            metadata: self.metadata,
-            frm_configs: None,
-            frm_config: self.frm_configs,
-            business_country: self.business_country,
-            business_label: self.business_label,
-            connector_label: self.connector_label,
-            business_sub_label: self.business_sub_label,
-            created_at: now,
-            modified_at: now,
-            connector_webhook_details: self.connector_webhook_details,
-            profile_id: Some(self.profile_id),
-            applepay_verified_domains: self.applepay_verified_domains,
-            pm_auth_config: self.pm_auth_config,
-            status: self.status,
-            connector_wallets_details: self.connector_wallets_details.map(Encryption::from),
-            additional_merchant_data: self.additional_merchant_data.map(|data| data.into()),
-            version: self.version,
-        })
-    }
-}
-
-#[cfg(feature = "v2")]
-#[async_trait::async_trait]
-impl behaviour::Conversion for MerchantConnectorAccount {
-    type DstType = storage::MerchantConnectorAccount;
-    type NewDstType = storage::MerchantConnectorAccountNew;
-
-    async fn convert(self) -> CustomResult<Self::DstType, ValidationError> {
-        Ok(storage::MerchantConnectorAccount {
-            id: self.id,
-            merchant_id: self.merchant_id,
-            connector_name: self.connector_name,
-            connector_account_details: self.connector_account_details.into(),
-            disabled: self.disabled,
-            payment_methods_enabled: self.payment_methods_enabled,
-            connector_type: self.connector_type,
-            metadata: self.metadata,
-            frm_config: self.frm_configs,
-            connector_label: self.connector_label,
-            created_at: self.created_at,
-            modified_at: self.modified_at,
-            connector_webhook_details: self.connector_webhook_details,
-            profile_id: self.profile_id,
-            applepay_verified_domains: self.applepay_verified_domains,
-            pm_auth_config: self.pm_auth_config,
-            status: self.status,
-            connector_wallets_details: self.connector_wallets_details.map(Encryption::from),
-            additional_merchant_data: self.additional_merchant_data.map(|data| data.into()),
-            version: self.version,
-            feature_metadata: self.feature_metadata.map(From::from),
-            connector_webhook_registration_details: None,
-        })
-    }
-
-    async fn convert_back(
-        state: &KeyManagerState,
-        other: Self::DstType,
-        key: &Secret<Vec<u8>>,
-        _key_manager_identifier: Identifier,
-    ) -> CustomResult<Self, ValidationError> {
-        let identifier = Identifier::Merchant(other.merchant_id.clone());
-
-        let decrypted_data = crypto_operation(
-            state,
-            type_name!(Self::DstType),
-            CryptoOperation::BatchDecrypt(EncryptedMerchantConnectorAccount::to_encryptable(
-                EncryptedMerchantConnectorAccount {
-                    connector_account_details: other.connector_account_details,
-                    additional_merchant_data: other.additional_merchant_data,
-                    connector_wallets_details: other.connector_wallets_details,
-                },
-            )),
-            identifier.clone(),
-            key.peek(),
-        )
-        .await
-        .and_then(|val| val.try_into_batchoperation())
-        .change_context(ValidationError::InvalidValue {
-            message: "Failed while decrypting connector account details".to_string(),
-        })?;
-
-        let decrypted_data = EncryptedMerchantConnectorAccount::from_encryptable(decrypted_data)
-            .change_context(ValidationError::InvalidValue {
-                message: "Failed while decrypting connector account details".to_string(),
-            })?;
-
-        Ok(Self {
-            id: other.id,
-            merchant_id: other.merchant_id,
-            connector_name: other.connector_name,
-            connector_account_details: decrypted_data.connector_account_details,
-            disabled: other.disabled,
-            payment_methods_enabled: other.payment_methods_enabled,
-            connector_type: other.connector_type,
-            metadata: other.metadata,
-
-            frm_configs: other.frm_config,
-            connector_label: other.connector_label,
-            created_at: other.created_at,
-            modified_at: other.modified_at,
-            connector_webhook_details: other.connector_webhook_details,
-            profile_id: other.profile_id,
-            applepay_verified_domains: other.applepay_verified_domains,
-            pm_auth_config: other.pm_auth_config,
-            status: other.status,
-            connector_wallets_details: decrypted_data.connector_wallets_details,
-            additional_merchant_data: decrypted_data.additional_merchant_data,
-            version: other.version,
-            feature_metadata: other.feature_metadata.map(From::from),
-        })
-    }
-
-    async fn construct_new(self) -> CustomResult<Self::NewDstType, ValidationError> {
-        let now = date_time::now();
-        Ok(Self::NewDstType {
-            id: self.id,
-            merchant_id: Some(self.merchant_id),
-            connector_name: Some(self.connector_name),
-            connector_account_details: Some(self.connector_account_details.into()),
-            disabled: self.disabled,
-            payment_methods_enabled: self.payment_methods_enabled,
-            connector_type: Some(self.connector_type),
-            metadata: self.metadata,
-            frm_config: self.frm_configs,
-            connector_label: self.connector_label,
-            created_at: now,
-            modified_at: now,
-            connector_webhook_details: self.connector_webhook_details,
-            profile_id: self.profile_id,
-            applepay_verified_domains: self.applepay_verified_domains,
-            pm_auth_config: self.pm_auth_config,
-            status: self.status,
-            connector_wallets_details: self.connector_wallets_details.map(Encryption::from),
-            additional_merchant_data: self.additional_merchant_data.map(|data| data.into()),
-            version: self.version,
-            feature_metadata: self.feature_metadata.map(From::from),
-        })
-    }
-}
-
-#[cfg(feature = "v1")]
-impl From<MerchantConnectorAccountUpdate> for MerchantConnectorAccountUpdateInternal {
-    fn from(merchant_connector_account_update: MerchantConnectorAccountUpdate) -> Self {
-        match merchant_connector_account_update {
-            MerchantConnectorAccountUpdate::Update {
-                connector_type,
-                connector_name,
-                connector_account_details,
-                test_mode,
-                disabled,
-                merchant_connector_id,
-                payment_methods_enabled,
-                metadata,
-                frm_configs,
-                connector_webhook_details,
-                applepay_verified_domains,
-                pm_auth_config,
-                connector_label,
-                status,
-                connector_wallets_details,
-                additional_merchant_data,
-            } => Self {
-                connector_type,
-                connector_name,
-                connector_account_details: connector_account_details.map(Encryption::from),
-                test_mode,
-                disabled,
-                merchant_connector_id,
-                payment_methods_enabled,
-                metadata,
-                frm_configs: None,
-                frm_config: frm_configs,
-                modified_at: Some(date_time::now()),
-                connector_webhook_details: *connector_webhook_details,
-                applepay_verified_domains,
-                pm_auth_config: *pm_auth_config,
-                connector_label,
-                status,
-                connector_wallets_details: connector_wallets_details.map(Encryption::from),
-                additional_merchant_data: additional_merchant_data.map(Encryption::from),
-                connector_webhook_registration_details: None,
-            },
-            MerchantConnectorAccountUpdate::ConnectorWalletDetailsUpdate {
-                connector_wallets_details,
-            } => Self {
-                connector_wallets_details: Some(Encryption::from(connector_wallets_details)),
-                connector_type: None,
-                connector_name: None,
-                connector_account_details: None,
-                connector_label: None,
-                test_mode: None,
-                disabled: None,
-                merchant_connector_id: None,
-                payment_methods_enabled: None,
-                frm_configs: None,
-                metadata: None,
-                modified_at: None,
-                connector_webhook_details: None,
-                frm_config: None,
-                applepay_verified_domains: None,
-                pm_auth_config: None,
-                status: None,
-                additional_merchant_data: None,
-                connector_webhook_registration_details: None,
-            },
-            MerchantConnectorAccountUpdate::ConnectorWebhookRegisterationUpdate {
-                connector_webhook_registration_details,
-            } => Self {
-                connector_type: None,
-                connector_name: None,
-                connector_account_details: None,
-                connector_label: None,
-                test_mode: None,
-                disabled: None,
-                merchant_connector_id: None,
-                payment_methods_enabled: None,
-                frm_configs: None,
-                metadata: None,
-                modified_at: None,
-                connector_webhook_details: None,
-                frm_config: None,
-                applepay_verified_domains: None,
-                pm_auth_config: None,
-                status: None,
-                connector_wallets_details: None,
-                additional_merchant_data: None,
-                connector_webhook_registration_details,
-            },
-        }
-    }
-}
-
-#[cfg(feature = "v2")]
-impl From<MerchantConnectorAccountUpdate> for MerchantConnectorAccountUpdateInternal {
-    fn from(merchant_connector_account_update: MerchantConnectorAccountUpdate) -> Self {
-        match merchant_connector_account_update {
-            MerchantConnectorAccountUpdate::Update {
-                connector_type,
-                connector_account_details,
-                disabled,
-                payment_methods_enabled,
-                metadata,
-                frm_configs,
-                connector_webhook_details,
-                applepay_verified_domains,
-                pm_auth_config,
-                connector_label,
-                status,
-                connector_wallets_details,
-                additional_merchant_data,
-                feature_metadata,
-            } => Self {
-                connector_type,
-                connector_account_details: connector_account_details.map(Encryption::from),
-                disabled,
-                payment_methods_enabled,
-                metadata,
-                frm_config: frm_configs,
-                modified_at: Some(date_time::now()),
-                connector_webhook_details: *connector_webhook_details,
-                applepay_verified_domains,
-                pm_auth_config: *pm_auth_config,
-                connector_label,
-                status,
-                connector_wallets_details: connector_wallets_details.map(Encryption::from),
-                additional_merchant_data: additional_merchant_data.map(Encryption::from),
-                feature_metadata: feature_metadata.map(From::from),
-            },
-            MerchantConnectorAccountUpdate::ConnectorWalletDetailsUpdate {
-                connector_wallets_details,
-            } => Self {
-                connector_wallets_details: Some(Encryption::from(connector_wallets_details)),
-                connector_type: None,
-                connector_account_details: None,
-                connector_label: None,
-                disabled: None,
-                payment_methods_enabled: None,
-                metadata: None,
-                modified_at: None,
-                connector_webhook_details: None,
-                frm_config: None,
-                applepay_verified_domains: None,
-                pm_auth_config: None,
-                status: None,
-                additional_merchant_data: None,
-                feature_metadata: None,
-            },
-        }
-    }
-}
-
 common_utils::create_list_wrapper!(
     MerchantConnectorAccounts,
     MerchantConnectorAccount,
@@ -1061,11 +647,10 @@ impl From<DieselMerchantConnectorAccountFeatureMetadata>
 
 #[async_trait::async_trait]
 pub trait MerchantConnectorAccountInterface
-where
-    MerchantConnectorAccount: behaviour::Conversion<
-        DstType = storage::MerchantConnectorAccount,
-        NewDstType = storage::MerchantConnectorAccountNew,
-    >,
+// MerchantConnectorAccount: behaviour::Conversion<
+//     DstType = storage::MerchantConnectorAccount,
+//     NewDstType = storage::MerchantConnectorAccountNew,
+// >,
 {
     type Error;
     #[cfg(feature = "v1")]
