+++ crates/storage_impl/src/merchant_connector_account.rs
@@ -3,7 +3,6 @@ use common_utils::{encryption::Encryption, ext_traits::AsyncExt};
 use diesel_models::merchant_connector_account as storage;
 use error_stack::{report, ResultExt};
 use hyperswitch_domain_models::{
-    behaviour::{Conversion, ReverseConversion},
     merchant_connector_account::{self as domain, MerchantConnectorAccountInterface},
     merchant_key_store::MerchantKeyStore,
 };
@@ -12,6 +11,7 @@ use router_env::{instrument, tracing};
 #[cfg(feature = "accounts_cache")]
 use crate::redis::cache;
 use crate::{
+    behaviour::{Conversion, ForeignFrom, ReverseConversion},
     kv_router_store,
     utils::{pg_accounts_connection_read, pg_accounts_connection_write},
     CustomResult, DatabaseStore, MockDb, RouterStore, StorageError,
@@ -464,9 +464,7 @@ impl<T: DatabaseStore> MerchantConnectorAccountInterface for RouterStore<T> {
                 self.get_keymanager_state()
                     .attach_printable("Missing KeyManagerState")?,
                 key_store.key.get_inner(),
-                common_utils::types::keymanager::Identifier::Merchant(
-                    key_store.merchant_id.clone(),
-                ),
+                keymanager::Identifier::Merchant(key_store.merchant_id.clone()),
             )
             .await
             .change_context(Self::Error::DecryptionError)
@@ -818,9 +816,7 @@ impl<T: DatabaseStore> MerchantConnectorAccountInterface for RouterStore<T> {
                         self.get_keymanager_state()
                             .attach_printable("Missing KeyManagerState")?,
                         key_store.key.get_inner(),
-                        common_utils::types::keymanager::Identifier::Merchant(
-                            key_store.merchant_id.clone(),
-                        ),
+                        keymanager::Identifier::Merchant(key_store.merchant_id.clone()),
                     )
                     .await
                     .change_context(Self::Error::DecryptionError)
@@ -1202,9 +1198,7 @@ impl MerchantConnectorAccountInterface for MockDb {
                         self.get_keymanager_state()
                             .attach_printable("Missing KeyManagerState")?,
                         key_store.key.get_inner(),
-                        common_utils::types::keymanager::Identifier::Merchant(
-                            key_store.merchant_id.clone(),
-                        ),
+                        keymanager::Identifier::Merchant(key_store.merchant_id.clone()),
                     )
                     .await
                     .change_context(StorageError::DecryptionError)
@@ -1306,9 +1300,7 @@ impl MerchantConnectorAccountInterface for MockDb {
                 self.get_keymanager_state()
                     .attach_printable("Missing KeyManagerState")?,
                 key_store.key.get_inner(),
-                common_utils::types::keymanager::Identifier::Merchant(
-                    key_store.merchant_id.clone(),
-                ),
+                keymanager::Identifier::Merchant(key_store.merchant_id.clone()),
             )
             .await
             .change_context(StorageError::DecryptionError)
@@ -1454,9 +1446,7 @@ impl MerchantConnectorAccountInterface for MockDb {
                         self.get_keymanager_state()
                             .attach_printable("Missing KeyManagerState")?,
                         key_store.key.get_inner(),
-                        common_utils::types::keymanager::Identifier::Merchant(
-                            key_store.merchant_id.clone(),
-                        ),
+                        keymanager::Identifier::Merchant(key_store.merchant_id.clone()),
                     )
                     .await
                     .change_context(StorageError::DecryptionError)
@@ -1515,6 +1505,445 @@ impl MerchantConnectorAccountInterface for MockDb {
                 )
                 .into())
             }
+        }
+    }
+}
+
+#[cfg(feature = "v2")]
+use std::collections::HashMap;
+
+use common_utils::{
+    errors::ValidationError,
+    type_name,
+    types::keymanager::{self, KeyManagerState, ToEncryptable},
+};
+#[cfg(feature = "v2")]
+use diesel_models::merchant_connector_account::{
+    BillingAccountReference as DieselBillingAccountReference,
+    MerchantConnectorAccountFeatureMetadata as DieselMerchantConnectorAccountFeatureMetadata,
+    RevenueRecoveryMetadata as DieselRevenueRecoveryMetadata,
+};
+#[cfg(feature = "v2")]
+use hyperswitch_domain_models::errors::api_error_response;
+use hyperswitch_domain_models::type_encryption::{crypto_operation, CryptoOperation};
+use hyperswitch_masking::{PeekInterface, Secret};
+
+#[cfg(feature = "v1")]
+#[async_trait::async_trait]
+impl Conversion for domain::MerchantConnectorAccount {
+    type DstType = storage::MerchantConnectorAccount;
+    type NewDstType = storage::MerchantConnectorAccountNew;
+
+    async fn convert(self) -> CustomResult<Self::DstType, ValidationError> {
+        Ok(storage::MerchantConnectorAccount {
+            merchant_id: self.merchant_id,
+            connector_name: self.connector_name,
+            connector_account_details: self.connector_account_details.into(),
+            test_mode: self.test_mode,
+            disabled: self.disabled,
+            merchant_connector_id: self.merchant_connector_id.clone(),
+            id: Some(self.merchant_connector_id),
+            payment_methods_enabled: self.payment_methods_enabled,
+            connector_type: self.connector_type,
+            metadata: self.metadata,
+            frm_configs: None,
+            frm_config: self.frm_configs,
+            business_country: self.business_country,
+            business_label: self.business_label,
+            connector_label: self.connector_label,
+            business_sub_label: self.business_sub_label,
+            created_at: self.created_at,
+            modified_at: self.modified_at,
+            connector_webhook_details: self.connector_webhook_details,
+            profile_id: Some(self.profile_id),
+            applepay_verified_domains: self.applepay_verified_domains,
+            pm_auth_config: self.pm_auth_config,
+            status: self.status,
+            connector_wallets_details: self.connector_wallets_details.map(Encryption::from),
+            additional_merchant_data: self.additional_merchant_data.map(|data| data.into()),
+            version: self.version,
+            connector_webhook_registration_details: self.connector_webhook_registration_details,
+        })
+    }
+
+    async fn convert_back(
+        state: &KeyManagerState,
+        other: Self::DstType,
+        key: &Secret<Vec<u8>>,
+        _key_manager_identifier: keymanager::Identifier,
+    ) -> CustomResult<Self, ValidationError> {
+        let identifier = keymanager::Identifier::Merchant(other.merchant_id.clone());
+        let decrypted_data = crypto_operation(
+            state,
+            type_name!(Self::DstType),
+            CryptoOperation::BatchDecrypt(
+                domain::EncryptedMerchantConnectorAccount::to_encryptable(
+                    domain::EncryptedMerchantConnectorAccount {
+                        connector_account_details: other.connector_account_details,
+                        additional_merchant_data: other.additional_merchant_data,
+                        connector_wallets_details: other.connector_wallets_details,
+                    },
+                ),
+            ),
+            identifier.clone(),
+            key.peek(),
+        )
+        .await
+        .and_then(|val| val.try_into_batchoperation())
+        .change_context(ValidationError::InvalidValue {
+            message: "Failed while decrypting connector account details".to_string(),
+        })?;
+
+        let decrypted_data =
+            domain::EncryptedMerchantConnectorAccount::from_encryptable(decrypted_data)
+                .change_context(ValidationError::InvalidValue {
+                    message: "Failed while decrypting connector account details".to_string(),
+                })?;
+
+        Ok(Self {
+            merchant_id: other.merchant_id,
+            connector_name: other.connector_name,
+            connector_account_details: decrypted_data.connector_account_details,
+            test_mode: other.test_mode,
+            disabled: other.disabled,
+            merchant_connector_id: other.merchant_connector_id,
+            payment_methods_enabled: other.payment_methods_enabled,
+            connector_type: other.connector_type,
+            metadata: other.metadata,
+
+            frm_configs: other.frm_config,
+            business_country: other.business_country,
+            business_label: other.business_label,
+            connector_label: other.connector_label,
+            business_sub_label: other.business_sub_label,
+            created_at: other.created_at,
+            modified_at: other.modified_at,
+            connector_webhook_details: other.connector_webhook_details,
+            profile_id: other
+                .profile_id
+                .ok_or(ValidationError::MissingRequiredField {
+                    field_name: "profile_id".to_string(),
+                })?,
+            applepay_verified_domains: other.applepay_verified_domains,
+            pm_auth_config: other.pm_auth_config,
+            status: other.status,
+            connector_wallets_details: decrypted_data.connector_wallets_details,
+            additional_merchant_data: decrypted_data.additional_merchant_data,
+            version: other.version,
+            connector_webhook_registration_details: other.connector_webhook_registration_details,
+        })
+    }
+
+    async fn construct_new(self) -> CustomResult<Self::NewDstType, ValidationError> {
+        let now = common_utils::date_time::now();
+        Ok(Self::NewDstType {
+            merchant_id: Some(self.merchant_id),
+            connector_name: Some(self.connector_name),
+            connector_account_details: Some(self.connector_account_details.into()),
+            test_mode: self.test_mode,
+            disabled: self.disabled,
+            merchant_connector_id: self.merchant_connector_id.clone(),
+            id: Some(self.merchant_connector_id),
+            payment_methods_enabled: self.payment_methods_enabled,
+            connector_type: Some(self.connector_type),
+            metadata: self.metadata,
+            frm_configs: None,
+            frm_config: self.frm_configs,
+            business_country: self.business_country,
+            business_label: self.business_label,
+            connector_label: self.connector_label,
+            business_sub_label: self.business_sub_label,
+            created_at: now,
+            modified_at: now,
+            connector_webhook_details: self.connector_webhook_details,
+            profile_id: Some(self.profile_id),
+            applepay_verified_domains: self.applepay_verified_domains,
+            pm_auth_config: self.pm_auth_config,
+            status: self.status,
+            connector_wallets_details: self.connector_wallets_details.map(Encryption::from),
+            additional_merchant_data: self.additional_merchant_data.map(|data| data.into()),
+            version: self.version,
+        })
+    }
+}
+
+#[cfg(feature = "v2")]
+#[async_trait::async_trait]
+impl Conversion for domain::MerchantConnectorAccount {
+    type DstType = storage::MerchantConnectorAccount;
+    type NewDstType = storage::MerchantConnectorAccountNew;
+
+    async fn convert(self) -> CustomResult<Self::DstType, ValidationError> {
+        Ok(storage::MerchantConnectorAccount {
+            id: self.id,
+            merchant_id: self.merchant_id,
+            connector_name: self.connector_name,
+            connector_account_details: self.connector_account_details.into(),
+            disabled: self.disabled,
+            payment_methods_enabled: self.payment_methods_enabled,
+            connector_type: self.connector_type,
+            metadata: self.metadata,
+            frm_config: self.frm_configs,
+            connector_label: self.connector_label,
+            created_at: self.created_at,
+            modified_at: self.modified_at,
+            connector_webhook_details: self.connector_webhook_details,
+            profile_id: self.profile_id,
+            applepay_verified_domains: self.applepay_verified_domains,
+            pm_auth_config: self.pm_auth_config,
+            status: self.status,
+            connector_wallets_details: self.connector_wallets_details.map(Encryption::from),
+            additional_merchant_data: self.additional_merchant_data.map(|data| data.into()),
+            version: self.version,
+            feature_metadata: self.feature_metadata.map(From::from),
+            connector_webhook_registration_details: None,
+        })
+    }
+
+    async fn convert_back(
+        state: &KeyManagerState,
+        other: Self::DstType,
+        key: &Secret<Vec<u8>>,
+        _key_manager_identifier: keymanager::Identifier,
+    ) -> CustomResult<Self, ValidationError> {
+        let identifier = keymanager::Identifier::Merchant(other.merchant_id.clone());
+
+        let decrypted_data = crypto_operation(
+            state,
+            type_name!(Self::DstType),
+            CryptoOperation::BatchDecrypt(
+                domain::EncryptedMerchantConnectorAccount::to_encryptable(
+                    domain::EncryptedMerchantConnectorAccount {
+                        connector_account_details: other.connector_account_details,
+                        additional_merchant_data: other.additional_merchant_data,
+                        connector_wallets_details: other.connector_wallets_details,
+                    },
+                ),
+            ),
+            identifier.clone(),
+            key.peek(),
+        )
+        .await
+        .and_then(|val| val.try_into_batchoperation())
+        .change_context(ValidationError::InvalidValue {
+            message: "Failed while decrypting connector account details".to_string(),
+        })?;
+
+        let decrypted_data =
+            domain::EncryptedMerchantConnectorAccount::from_encryptable(decrypted_data)
+                .change_context(ValidationError::InvalidValue {
+                    message: "Failed while decrypting connector account details".to_string(),
+                })?;
+
+        Ok(Self {
+            id: other.id,
+            merchant_id: other.merchant_id,
+            connector_name: other.connector_name,
+            connector_account_details: decrypted_data.connector_account_details,
+            disabled: other.disabled,
+            payment_methods_enabled: other.payment_methods_enabled,
+            connector_type: other.connector_type,
+            metadata: other.metadata,
+
+            frm_configs: other.frm_config,
+            connector_label: other.connector_label,
+            created_at: other.created_at,
+            modified_at: other.modified_at,
+            connector_webhook_details: other.connector_webhook_details,
+            profile_id: other.profile_id,
+            applepay_verified_domains: other.applepay_verified_domains,
+            pm_auth_config: other.pm_auth_config,
+            status: other.status,
+            connector_wallets_details: decrypted_data.connector_wallets_details,
+            additional_merchant_data: decrypted_data.additional_merchant_data,
+            version: other.version,
+            feature_metadata: other.feature_metadata.map(From::from),
+        })
+    }
+
+    async fn construct_new(self) -> CustomResult<Self::NewDstType, ValidationError> {
+        let now = common_utils::date_time::now();
+        Ok(Self::NewDstType {
+            id: self.id,
+            merchant_id: Some(self.merchant_id),
+            connector_name: Some(self.connector_name),
+            connector_account_details: Some(self.connector_account_details.into()),
+            disabled: self.disabled,
+            payment_methods_enabled: self.payment_methods_enabled,
+            connector_type: Some(self.connector_type),
+            metadata: self.metadata,
+            frm_config: self.frm_configs,
+            connector_label: self.connector_label,
+            created_at: now,
+            modified_at: now,
+            connector_webhook_details: self.connector_webhook_details,
+            profile_id: self.profile_id,
+            applepay_verified_domains: self.applepay_verified_domains,
+            pm_auth_config: self.pm_auth_config,
+            status: self.status,
+            connector_wallets_details: self.connector_wallets_details.map(Encryption::from),
+            additional_merchant_data: self.additional_merchant_data.map(|data| data.into()),
+            version: self.version,
+            feature_metadata: self.feature_metadata.map(From::from),
+        })
+    }
+}
+
+#[cfg(feature = "v1")]
+impl ForeignFrom<domain::MerchantConnectorAccountUpdate>
+    for storage::MerchantConnectorAccountUpdateInternal
+{
+    fn foreign_from(
+        merchant_connector_account_update: domain::MerchantConnectorAccountUpdate,
+    ) -> Self {
+        match merchant_connector_account_update {
+            domain::MerchantConnectorAccountUpdate::Update {
+                connector_type,
+                connector_name,
+                connector_account_details,
+                test_mode,
+                disabled,
+                merchant_connector_id,
+                payment_methods_enabled,
+                metadata,
+                frm_configs,
+                connector_webhook_details,
+                applepay_verified_domains,
+                pm_auth_config,
+                connector_label,
+                status,
+                connector_wallets_details,
+                additional_merchant_data,
+            } => Self {
+                connector_type,
+                connector_name,
+                connector_account_details: connector_account_details.map(Encryption::from),
+                test_mode,
+                disabled,
+                merchant_connector_id,
+                payment_methods_enabled,
+                metadata,
+                frm_configs: None,
+                frm_config: frm_configs,
+                modified_at: Some(common_utils::date_time::now()),
+                connector_webhook_details: *connector_webhook_details,
+                applepay_verified_domains,
+                pm_auth_config: *pm_auth_config,
+                connector_label,
+                status,
+                connector_wallets_details: connector_wallets_details.map(Encryption::from),
+                additional_merchant_data: additional_merchant_data.map(Encryption::from),
+                connector_webhook_registration_details: None,
+            },
+            domain::MerchantConnectorAccountUpdate::ConnectorWalletDetailsUpdate {
+                connector_wallets_details,
+            } => Self {
+                connector_wallets_details: Some(Encryption::from(connector_wallets_details)),
+                connector_type: None,
+                connector_name: None,
+                connector_account_details: None,
+                connector_label: None,
+                test_mode: None,
+                disabled: None,
+                merchant_connector_id: None,
+                payment_methods_enabled: None,
+                frm_configs: None,
+                metadata: None,
+                modified_at: None,
+                connector_webhook_details: None,
+                frm_config: None,
+                applepay_verified_domains: None,
+                pm_auth_config: None,
+                status: None,
+                additional_merchant_data: None,
+                connector_webhook_registration_details: None,
+            },
+            domain::MerchantConnectorAccountUpdate::ConnectorWebhookRegisterationUpdate {
+                connector_webhook_registration_details,
+            } => Self {
+                connector_type: None,
+                connector_name: None,
+                connector_account_details: None,
+                connector_label: None,
+                test_mode: None,
+                disabled: None,
+                merchant_connector_id: None,
+                payment_methods_enabled: None,
+                frm_configs: None,
+                metadata: None,
+                modified_at: None,
+                connector_webhook_details: None,
+                frm_config: None,
+                applepay_verified_domains: None,
+                pm_auth_config: None,
+                status: None,
+                connector_wallets_details: None,
+                additional_merchant_data: None,
+                connector_webhook_registration_details,
+            },
+        }
+    }
+}
+
+#[cfg(feature = "v2")]
+impl ForeignFrom<domain::MerchantConnectorAccountUpdate>
+    for storage::MerchantConnectorAccountUpdateInternal
+{
+    fn foreign_from(
+        merchant_connector_account_update: domain::MerchantConnectorAccountUpdate,
+    ) -> Self {
+        match merchant_connector_account_update {
+            domain::MerchantConnectorAccountUpdate::Update {
+                connector_type,
+                connector_account_details,
+                disabled,
+                payment_methods_enabled,
+                metadata,
+                frm_configs,
+                connector_webhook_details,
+                applepay_verified_domains,
+                pm_auth_config,
+                connector_label,
+                status,
+                connector_wallets_details,
+                additional_merchant_data,
+                feature_metadata,
+            } => Self {
+                connector_type,
+                connector_account_details: connector_account_details.map(Encryption::from),
+                disabled,
+                payment_methods_enabled,
+                metadata,
+                frm_config: frm_configs,
+                modified_at: Some(common_utils::date_time::now()),
+                connector_webhook_details: *connector_webhook_details,
+                applepay_verified_domains,
+                pm_auth_config: *pm_auth_config,
+                connector_label,
+                status,
+                connector_wallets_details: connector_wallets_details.map(Encryption::from),
+                additional_merchant_data: additional_merchant_data.map(Encryption::from),
+                feature_metadata: feature_metadata.map(From::from),
+            },
+            domain::MerchantConnectorAccountUpdate::ConnectorWalletDetailsUpdate {
+                connector_wallets_details,
+            } => Self {
+                connector_wallets_details: Some(Encryption::from(connector_wallets_details)),
+                connector_type: None,
+                connector_account_details: None,
+                connector_label: None,
+                disabled: None,
+                payment_methods_enabled: None,
+                metadata: None,
+                modified_at: None,
+                connector_webhook_details: None,
+                frm_config: None,
+                applepay_verified_domains: None,
+                pm_auth_config: None,
+                status: None,
+                additional_merchant_data: None,
+                feature_metadata: None,
+            },
         }
     }
 }
